requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==23.0.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.4
python-multipart==0.0.20
